+++ tagger/src/components/mod.rs
@@ -9,7 +9,7 @@ mod title;
 pub(crate) use checkbox::CheckBox;
 pub(crate) use footer::{CmperFooter, PickerFooter, TaggerFooter};
 pub(crate) use histogram::Histogram;
-pub(crate) use image::{Grid, Image, Images};
+pub(crate) use image::{preload, Grid, Image, Images};
 pub(crate) use input::Input;
 pub(crate) use quit::Quit;
 pub(crate) use title::Title;